added folder separations added storage manager protocol and its implementation
